+++ Logger/Logger.php
@@ -5,8 +5,9 @@ namespace Conekta\Payments\Logger;
 use Conekta\Payments\Helper\Data;
 use Magento\Framework\App\ObjectManager;
 use Monolog\DateTimeImmutable;
+use Monolog\Logger;
 
-class Logger extends \Monolog\Logger
+class Logger
 {
     /**
      * @param int $level
@@ -14,13 +15,13 @@ class Logger extends \Monolog\Logger
      * @param array $context
      * @return bool
      */
-    public function addRecord(int $level, string $message, array $context = [], ?Monolog\DateTimeImmutable $datetime = null): bool
+    public function addRecord(int $level, string $message, array $context = []): bool
     {
         $objectManager = ObjectManager::getInstance();
         $conektaHelper = $objectManager->create(Data::class);
 
         if ((int)$conektaHelper->getConfigData('conekta/conekta_global', 'debug')) {
-            return parent::addRecord($level, $message, $context);
+            return Logger::addRecord($level, $message, $context);
         }
 
         return true;